requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==2.0.0
python-dotenv==1.0.1
email-validator==2.1.1
Flask-Bcrypt==1.0.1
